refactor: Sonar 이슈 수정
